Not found integration tests

+++ sequalize-way/controllers/UsersController.js
@@ -10,7 +10,7 @@ export default class UsersController {
     const validationErrors = validateUser(req.body);
     if (validationErrors) {
       return res.status(HttpStatus.BAD_REQUEST).send(validationErrors);
-    }
+    }         
     const newUser = await this.service.create(req.body);
     res.status(HttpStatus.CREATED);
     res.send(newUser);
@@ -21,21 +21,34 @@ export default class UsersController {
     res.status(HttpStatus.OK).send(users);
   };
 
-  update = async (req, res) => {
+  update = async (req, res, next) => {
     const validationErrors = validateUser(req.body);
     if (validationErrors) {
       return res.status(HttpStatus.BAD_REQUEST).send(validationErrors);
     }
+
+    const existingUser = await this.service.getById(req.params.id);
+    if (!existingUser) {
+      return res.sendStatus(HttpStatus.NOT_FOUND);
+    }
+
     await this.service.update(req.body, req.params.id);
     res.sendStatus(HttpStatus.NO_CONTENT);
   };
 
   getById = async (req, res) => {
     const existingUser = await this.service.getById(req.params.id);
+    if (!existingUser) {
+      return res.sendStatus(HttpStatus.NOT_FOUND);
+    }
     res.status(HttpStatus.OK).send(existingUser);
   };
 
   destroy = async (req, res) => {
+    const existingUser = await this.service.getById(req.params.id);
+    if (!existingUser) {
+      return res.sendStatus(HttpStatus.NOT_FOUND);
+    }
     await this.service.destroy(req.params.id);
     res.sendStatus(HttpStatus.NO_CONTENT);
   };
